training/synthetic_generation/requirements.txt
verovio>=4.0.0
pillow>=10.0.0
opencv-python>=4.8.0
lxml>=4.9.0
numpy>=1.24.0
scipy>=1.11.0
tqdm>=4.65.0
pyyaml>=6.0
